time_in_msec from Util should return int64
since the time returned is in msecs, the value will overflow a 32bit
number, so instead use 64 bit.

+++ Util.cpp
@@ -41,7 +41,7 @@ void sum(timeval &l, const timeval r)
 	l.tv_sec += r.tv_sec;
 }
 
-int time_in_msec(const timeval t)
+int64 time_in_msec(const timeval t)
 {
-	return t.tv_sec * 1000 + t.tv_usec / 1000;
+	return int64(t.tv_sec) * 1000 + int64(t.tv_usec) / 1000;
 }
